move AnimatedRoutes, Protected and Suspended into utility subfolder
