+++ client/src/App.tsx
@@ -1,4 +1,4 @@
-import AnimatedRoutes from "@/components/AnimatedRoutes";
+import AnimatedRoutes from "@/components/utility/AnimatedRoutes";
 import { createTheme, DEFAULT_THEME, MantineProvider } from "@mantine/core";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
